Added rsync options

+++ upload_queue_runner.plx
@@ -14,11 +14,18 @@ use warnings;
 
 use DBI;
 use Sys::Syslog qw/ :DEFAULT setlogsock /;
+use HTTP::Request;
+use LWP::UserAgent;
 
 require "./environ.pm";
 our $dbh;
 our $conf;
 our $gb_short_year;
+our $rsync_user;
+our $rsync_pass;
+our $rsync_path;
+our $rsync_host;
+
 my $sth;
 
 setlogsock('unix');
@@ -27,7 +34,7 @@ $ENV{PATH} = "/bin:/usr/bin";
 
 $SIG{ALRM} = sub {
 
-	openlog('gb_talks_upload_queue_runner','','user')
+	openlog('gb_talks_upload_queue_runner','','user');
 	syslog('err', 'Process took too long to complete - maybe rsync died?');
 	closelog;
 
@@ -57,27 +64,29 @@ if ( $current_uploads <= $max_uploads )
 	my $talk_id = pop @queue;
 	
 	alarm(3600); # Let the script run for an hour. If it takes longer than that, we want to quit, log any results, and let another process start up to resume the transfer. 
-	
-	$0 = "upload_queue_runner.plx - gb$gb_short_year-$talk_id.mp3";	
-	# Run the upload job
-	system("rsync --partial upload_queue/gb$gb_short_year-$talk_id.mp3 $rsync_user\@$rsync_host:$rsync_path/gb$gb_short_year-$talk_id.mp3");
+	if($talk_id)
+	{	
+		$0 = "upload_queue_runner.plx - gb$gb_short_year-$talk_id.mp3";	
+		
+		# Run the upload job
+		system("rsync --partial upload_queue/gb$gb_short_year-$talk_id.mp3 $rsync_user\@$rsync_host:$rsync_path/gb$gb_short_year-$talk_id.mp3");
 
-	# Check what return code rsync gave to determine how it did at the upload
-	
-	if ($? == -1) {
-		log('err', "failed to run rsync: $!\n");
-	}	
-	elsif ($? & 127) {
-		log(printf "child died with signal %d, %s coredump\n", ($? & 127), ($? & 128) ? "with" : "without");
-	}
-	else { # If we succeeded
-		log(printf "child exited with value %d\n", $? >> 8);
-		# Remove the item from the queue
-		$sth = $dbh->prepare('DELETE FROM upload_queue where talk_id=?');
+		# Check what return code rsync gave to determine how it did at the upload
+		if ($? == -1) {
+			log_it('err', "failed to run rsync: $!\n");
+		}	
+		elsif ($? & 127) {
+			log_it(printf "child died with signal %d, %s coredump\n", ($? & 127), ($? & 128) ? "with" : "without");
+		}
+		else { # If we succeeded
+			log_it(printf "child exited with value %d\n", $? >> 8);
+			# Remove the item from the queue
+			$sth = $dbh->prepare('DELETE FROM upload_queue where talk_id=?');
 		$sth->execute($talk_id);
-	}
+		}
 	
-	alarm(0);
+		alarm(0);
+	}
 }
 else
 {
@@ -87,10 +96,10 @@ else
 # Remove lockfile
 `rm /var/run/gb_upload$$`;
 
-sub log
+sub log_it
 {
 	my ($log_level, $log_message) = $@;
-	openlog('gb_talks_upload_queue_runner','','user')
+	openlog('gb_talks_upload_queue_runner','','user');
         syslog($log_level, $log_message);
         closelog;
 }
